Consolidate error handling into `Result` and `PollResult`
With the way error codes are handled in WolfSSL (where a non-`WOLFSSL_SUCCESS`
does not necessarily signify that something broke), I've opted to defer most of
the complexity to the future.

Right now we only really care for retry errors or fatal errors, so organize
errors along those lines.