Implement `trigger_update_keys`
Note that part of this changeset subtly alters the behavior of `try_read`, which
results in some changes to the 0 length case of the test
`try_read_and_write_roundtrip_once`.

I find this (I suppose existing wolfssl) behavior counterintuitive, but since
I'm expecting it to be very irregular for anyone to be passing in a buffer of
capacity 0 into `try_read` I'm willing to tolerate this edge behavior.