Plumb in an IOCallbacks trait through context and session
The an instance of the trait is mandatory since the callback model is
non-optional in these bindings.

The trait currently contains no methods, this is just the plumbing. In tests
dummy types are used.

We use generic types through out, rather than e.g. `Box<dyn IOCallbacks>` since
this API is used on the data path so we want to monomorphize down to concrete
types rather than do dynamic dispatch.