Split `Poll::Pending` into `Read` and `Write` variants.
This exposes what wolfSSL is waiting for to the application so that it can make
decide how to poll for I/O on the underlying transport.

This literally just splits the variant into two but makes no use of the
distinction yet. Applications should continue to process IO as they were before
treating `PendingRead` and `PendingWrite` identically.