Separate `try_shutdown`'s `WOLFSSL_SHUTDOWN_NOT_DONE` result from I/O
This error code does not indicate that any I/O is pending, it just indicates
that thus far the connection has only been unidirectionally shutdown, if the
application does not care about bidirectional shutdown (because it does not
intend to reuse the connection) then it can carry on.

If the application does care then it should call `try_shutdown` again without
e.g. polling for I/O availability on the underlying socket. If `try_shutdown`
determines that I/O is needed then it will return `WANT_{READ,WRITE}` as usual.