Mark `NonNull<wolfssl_sys::WOLFSSL>` and `WOLFSSL_CTX` as`Send`
This is safe according to the WolfSSL documentation.

The `NonNull` type has an explicit `impl<T: ?Sized> !Send for NonNull<T> {}`,
rather than inheriting from the properties of `T` (as some other wrapper types
do), so we cannot simply `unsafe impl Send for WOLFSSL` in the `wolfssl-sys`
crate and have `NonNull<WOLFSSL>` become `Send` automatically.

Instead with `NonNull` we must instead add a new-type wrapper and implement
`Send` on that.

For `WOLFSSL_CTX` they recommend only a single thread writes to the context,
which we already achieve by the use of the `ContextBuilder` (which is not
send). Once a `Context` is built we actually go further than my reading of the
documentation by synchronizing (via a `Mutex`) even _read_ access to the
context, when the documentation only requires synchronisation for _write_
access. Related to this (and because we don't want to expose `ContextPointer`
the `Context::ctx` method is only intended for use within this crate, so switch
to `pub(crate)`.