Remove `Mutex` from around `Context::ctx`
Once built the `Context` is intended to be used read only to initialize new
instances of `Session`. The WolfSSL documentation (see the comment on the
implementation of `Send` for `ContextPointer`) only requires synchronization
for write actions (which for us all happen on the `ContextBuilder` which is not
`Send`).

Therefore we do not need the `Mutex`, but as an extra layer of safety mark the
`Context::ctx` as `unsafe` and explain the read-only requirement. The function
is `pub(crate)` so there can be no external users.