Remove bindings for `wolfSSL_Cleanup`
This function deinitializes the library, meaning that all the safety guarentees
of other functions and types (including existing instances of those types)
become potentially invalid.

We could make the binding itself `unsafe` to force consumers to acknowledge
this, but since we currently do not have a use case for tearing down the
library (which is rather a niche desire anyway) lets just remove the bindings
until we have a reason to support this.

The tests no longer tear down the library, that's fine.