Use a `OnceLock` to ensure `wolfSSL_Init` is only called once.
The WolfSSL documentation says "must be called once per application", it is not
clear if this means "at most once" or "at least once". We take a conservative
stance at arrange for "at most once" behaviour.

The [implementation][] does actually call this for us in `wolfSSL_CTX_new_ex`
(used by `wolfSSL_CTX_new`) but we choose not to rely on this undocumented
implementation decision.

The documentation also says "and before any other call to the library", we
leave this up to the application for now.

We use a `OnceLock` so that we can repeatedly return the result for subsequent
calls, in order to support that our `Error` type must implement `Clone` so
derive it where needed.

[implementation]: https://github.com/wolfSSL/wolfssl/blob/v5.6.3-stable/src/ssl.c#L1271