Import `wolfssl-sys` crate
This imports precisely the content of the crate from
https://github.com/expressvpn/wolfssl-sys/commit/759881fc018627d999c8694c32b52fac4f941ad6
with no changes.

```console
$ git -C ../wolfssl-sys rev-parse HEAD
759881fc018627d999c8694c32b52fac4f941ad6
$ git -C ../wolfssl-sys status
On branch main
Your branch is up to date with 'origin/main'.

nothing to commit, working tree clean
$ cp -r ../wolfssl-sys/{Cargo.toml,build.rs,src,examples,wrapper.h} wolfssl-sys/
$ git add wolfssl-sys
```

Note that this is not all the files from the repository since many of them are
already covered here by workspace wide files. In particular the following files
are identical in both repositories: `CODE_OF_CONDUCT.adoc`, `CONTRIBUTING.adoc`,
`EXPRESSVPN_CLA.adoc`, `LICENSE` and `SECURITY.adoc`.

The `vendor` directory is deliberately omitted since we intend to replace it
with a git submodule.

These new files are not yet wired in to the cargo workspace and will need
tailoring before they are usable anyway.