LIT-122: GHA CI: rename cargo-deny job to check-dependencies
It's probably more useful for humans to understand what the
job does, not how the job implements the funtionality. It's also better
to decouple the naming here with the implementation detail should we
choose to achieve the license checking with a different tool in the
future.

Lastly, this makes the naming consistent with the "CI job name matches
Earthly target name" pattern.