ci: use repo secrets instead of environment secrets
There was a lot of noise in the PRs due to the environment secret usuage.
Everytime we made a change in the PR, the PR is flooded with

`deployed to environment...`

messages.

So we moved the secrets from environment to repo level. This commit removes
the environment value from ci so that github actions will use repo secrets.