Earthfile: Generate coverage results from actual rusts
... not whatever happens to be in the `target/` cache when we run `llvm-cov
report`.

There are two parts of this, first we must specify the coverage target
directory as outputs so `rust-udc+CARGO` will propagate them from the target
cache mount to the local target. Second we must not use the target cache when
we generate the reports.

The full `target/llvm-cov-target` does appear to be needed to generate the
reports.