CVPN-249 Pin wolfssl back to 0.0.1 to avoid a release
Crate.io requires all dependencies of a crate already present on
crate.io. Since wolfssl depends on wolfssl-sys, we need to release
wolfssl-sys first and modify wolfssl to depend on the wolfssl-sys crate,
instead of locally.

We could modify wolfssl to point to the wolfssl-sys 0.1.14 in this
change, but it would break because the 0.1.14 was released long time
ago, and does not contain any meaningful code. We can do that once a
newer version of wolfssl-sys has been released.

The GitHub Action detected a need to release the wolfssl crate based on
the Cargo.toml file. The latest version of wolfssl on crate.io is 0.0.1,
but it was set to 0.1.0 on the local crate file. To prevent a release,
we must change the version back to 0.0.1.

Since the crate has never been meaningfully released, it is safe to
revert it to a lower semver.