CVPN-249 Add version reference in wolfssl's Cargo definition
This change is to prepare for the wolfssl's release, which will be done
in future PRs.

Cargo requires alll dependencies of a crate points to a version on
crates.io during publish, so here a version reference is added.
According to the cargo docs [1], both the path and the version attribute
can be specified at the same time, and they will be used in different
scenarios.

path:    used during local development environment, and ci
version: used during publish

[1]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#specifying-path-dependencies