CVPN-249 Update Release Instruction
The cargo documentation [1] is slightly misleading. It was understood
that the version attribute will be ignored when not publishing, but
in fact it is not ignored. After re-reading the docs, its behaviour is:

During local/ci:
- Cargo will follow the path to locate the crate
- After the crate is located, the crate's version will be validated
  against the one specified

During publish:
- Cargo will only refer to the version in crates.io

Due to this updated understanding, we need to change our release
instruction to update both the `wolfssl-sys` crate version and the
version specified under the dependencies in the `wolfssl` crate.

[1]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html