GHA: Add workflow to validate wolfssl->wolfssl-sys dependency is up to date
When releasing wolfssl we want to be sure it depends on the most recent
wolfssl-sys changes.

For usage from git this is fine since the dependency involves a `path` key.

For usage from crates.io this means we want the `version` field to point to the
latest version of the code -- which may mean we need to release wolfssl-sys
first if there are changes there.

The SHA of the tree object of the `wolfssl-sys` subtree changes precisely when
wolfssl-sys is changed, so we can match using that.

This adds a workflow which checks both that wolfssl-sys is fully up to date and
that wolfssl depends on that version.