wolfssl: Use negotiated `ProtocolVersion` where needed
The `Protocol` type corresponds to a request to WolfSSL at setup time to take
on a particular role (client vs server) supporting some set of versions (DTLS,
TLS, 1.x etc). Accordingly it contains some "umbrella" versions such as
`DtlsClient` and `TlsServer` which request only "some version of" the
respective protocols.

Therefore it does not make sense to ask "is this protocol DTLS 1.3 etc" since
that would be false for `DtlsClient` even though the protocol which has
actually been negotiated may well be version 1.3.

Move the helper methods to the `ProtocolVersion` type which reflects the
runtime negotiated actual version and make use of `Session::version()` in the
places which previously used the `protocol` field (which is now unneeded).