wolfssl: Enable `missing_safety_doc` and `unnecessary_safety_comment` lints
This highlights two cases.

In one case, in `tls13_secret_cb`, the comment is relates to a subsequent
operation which is unsafe but which is currently unannotated because we are
within an `unsafe` function.  We will shortly be enabling
`unsafe_op_in_unsafe_fn` to enforce use of `unsafe` aroud such operations so do
so now and attach the comment to it.

In the other case, in `enable_tls13_keylog`, the comment really applies to an
operation within a subsequent `unsafe` block, move it there. We will shortly be
enabling `multiple_unsafe_ops_per_block` at which point things will be
clarified further.