windows: build wolfssl with msbuild
This change introduces a branch for building wolfssl in windows using
MsBuild. Before it can start building, we have to copy several override
files to customize our build options:

* wolfssl-user_settings-{pointer_width}
* options.h
* wolfssl.vcxproj

Particularly, options.h is handcrafted only on windows. That is because
MsBuild does not generate header files after build, which would lead to
the bindgen not knowing which option we have selected in
wolfssl.vcxproj.