CVPN-1846 Build WolfSSL on iOS
Build WolfSSL on the iOS platform with similar build flags from lightway-core. We would need to set `IPHONEOS_DEPLOYMENT_TARGET` in the environment variable as well in the cargo config, as it is needed to make sure we support older iOS version as well(now we are supporting iOS 15 or above).

Ref: https://doc.rust-lang.org/rustc/platform-support/apple-ios.html#os-version