wolfssl: "INIT_MP_INT_SIZE failed" fix on ARM32
According to the docs, the marco "SP_INT_BITS" determines the number of bits in sp_int, which is the largest bignum that can be handled.
https://www.wolfssl.com/documentation/manuals/wolfssl/chapter02.html#sp_int_bits

By default it is 3072. Since we are using 4096 bits keys, we need to add the c flag "-DSP_INT_BITS=4096" override the marco and support 4096 bits keys.