wolfssl: use proper warnings on MSVC
This error is found when we are attempting to upgrade lightway-core in C. This patch is added to save some trouble when we add support to Windows in the future.

One code line in wolfssl v5.7.6 used a preprocessor directive called "warning", which is not supported by MSVC compiler.

The follow error shows:
wolfssl/wolfcrypt/settings.h(355,1): fatal  error C1021: invalid preprocessor command 'warning'

To fix this, pragma shall be used instead for MSVC compiler.

credit: Raihaan Shouhell