CVPN-2045: Fix not compiling in risc-v using dev profile
This is needed to make wolfssl-sys compiles using dev profile, and is a workaround for CVPN-2045.
This is a workaround to stop compiler and inline asm used in wolfssl from contesting for register s0 when compiling using -o0, not -o3.
All test cases pass even after the change.
This flag is already being used in release mode before this change, since release profile causes the ./configure to use -o3, and -o3 automatically turns on -fomit-frame-pointer.
We may remove this workaround once an alternative solution is found, or wolfssl resolves the issue.