ci: use github.event.number for pullrequest
Until now, we use gh-find-current-pr to find the PR number of the
specified commit. This was needed when the action is triggered from
main branch push event, since we do not know the PR number from context.

This also works fine for pull requests from branches created in main repo
But it is failing for forked repos.

To overcome this issue, use github.event.number for pull request event and
use the one from above action if it is push event.