wolfssl-sys: fix connect_pq example
This test is broken for sometime and we did not notice it.
Since we move from Kyber to MLKEM, we need to update the test port
to use MLKEM based PQ connections.

With this change, clippy started failing with
```
+lint *failed* | error[E0425]: cannot find value `WOLFSSL_P521_ML_KEM_1024` in crate `ffi`
+lint *failed* |   --> wolfssl-sys/examples/connect_pq.rs:69:54
+lint *failed* |    |
+lint *failed* | 69 |         let res = ffi::wolfSSL_UseKeyShare(ssl, ffi::WOLFSSL_P521_ML_KEM_1024 as u16);
+lint *failed* |    |                                                      ^^^^^^^^^^^^^^^^^^^^^^^^ not found in `ffi`
```

Its because we are running clippy with all features, and since
kyber_only is one of the feature which disables MLKEM.

To fix this, run clippy twice once without examples and one with example.